move theme config files to design system folder
