+++ src/constants/index.ts
@@ -1,10 +1,8 @@
 export * from './asyncStorageKeys'
-export * from './breakpoints'
 export * from './colorSchemes'
 export * from './env'
 export * from './environments'
 export * from './images'
 export * from './navigation'
 export * from './regex'
-export * from './theme'
 export * from './time'
