+++ src/design-system/components/Button/Button.tsx
@@ -1,5 +1,3 @@
-import { theme } from '@baca/constants'
-import { buttonVariants } from '@baca/constants/buttonVariants'
 import { useColorScheme } from '@baca/contexts'
 import { getColorValue } from '@baca/utils'
 import {
@@ -22,6 +20,7 @@ import {
   View,
 } from 'react-native'
 
+import { buttonVariants, theme } from '../../config'
 import { generateStyledComponent } from '../../utils'
 import { Box } from '../Box'
 import { Loader } from '../Loader'
