+++ src/design-system/components/Text/Text.tsx
@@ -1,9 +1,9 @@
-import { textVariants } from '@baca/constants/textVariants'
 import { useTheme } from '@baca/hooks'
 import { getColorValue, convertEmToNumber } from '@baca/utils'
 import { useMemo, memo, forwardRef, PropsWithoutRef, RefAttributes } from 'react'
 import { TextProps as BaseTextProps, Text as BaseText, TextStyle } from 'react-native'
 
+import { textVariants } from '../../config'
 import { generateStyledComponent, generateStyleSheet } from '../../utils'
 import { StyledProps } from '../types'
