+++ src/design-system/config/index.ts
@@ -1,0 +1,5 @@
+export * from './breakpoints'
+export * from './buttonVariants'
+export * from './colors'
+export * from './textVariants'
+export * from './theme'
