+++ src/design-system/index.ts
@@ -1,3 +1,4 @@
 export * from './bottomSheets'
 export * from './components'
+export * from './config'
 export * from './utils'
