+++ src/hooks/useTheme.ts
@@ -1,5 +1,5 @@
-import { theme } from '@baca/constants'
 import { useColorScheme } from '@baca/contexts'
+import { theme } from '@baca/design-system/config'
 
 export const useTheme = (): AppTheme => {
   const { colorScheme } = useColorScheme()
